move CloseableConsumer to commons.lang
