+++ airbyte-commons/src/main/java/io/airbyte/commons/yaml/Yamls.java
@@ -29,7 +29,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SequenceWriter;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import io.airbyte.commons.closeable.CloseableConsumer;
+import io.airbyte.commons.lang.CloseableConsumer;
 import io.airbyte.commons.lang.Exceptions;
 import java.io.IOException;
 import java.io.Writer;
